Refactor:#321 - decouple sidebar control from NotesClient to NotesPage
